Leave `flip` handling out of TiledHeatmapMesh
